Add some macros useful when writing procedural syntax extensions
These are adapted [1] from html5ever, which is licensed MIT/ASL2.

I wasn't sure if you planned for people to use this library as

    #[macro_use]
    extern crate mac;

i.e. with no list of specific macros to import. For that reason I gave these
macros an "ext_" prefix, as they are unlikely to be useful outside of writing
syntax extensions.

[1]: https://github.com/servo/html5ever/blob/master/macros/src/internal.rs