Add possibility to use `break` and `continue` in `do_while!`
This is done using a variable that tracks whether the loop already
executed. This allows using `break` and `continue` in the body of the
loop. It would be preferable to express the do-while-loop as
```
loop {
    $body
    if $cond {
        break;
    }
}
```
however, this does not work well because of the interaction with
`continue`.

This is a breaking change due to interactions with initialization
tracking, because `rustc` can no longer see that the body is guaranteed
to execute.